Add disk sync after each check for AMI status
This should give us some extra logs while we are polling, so it doesn't just end after we call to build the AMI.
It might be this gives us nothing, or it might be better to actually just shutdown, as the instance might not terminate while the AMI is still generating, and might wait until the AMI is completed. More research will be required, but doing that might give us a more friendly log on a finished AMI, but might, unfortunately, give us less insight in the event something went wrong.